test: split shared test infra and delete helper monolith
