+++ tests/unit/cli/test_analytics.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 from polylogue.cli.analytics import ProviderMetrics, compute_provider_comparison
-from tests.infra.helpers import make_conversation, make_message
+from tests.infra.storage_records import make_conversation, make_message
 
 
 class TestProviderMetrics:
