+++ tests/unit/core/conftest.py
@@ -17,7 +17,7 @@ from polylogue.storage.backends.async_sqlite import SQLiteBackend
 from polylogue.storage.backends.connection import open_connection
 from polylogue.storage.index import rebuild_index
 from polylogue.storage.repository import ConversationRepository
-from tests.infra.helpers import ConversationBuilder
+from tests.infra.storage_records import ConversationBuilder
 
 
 @pytest.fixture
