+++ tests/unit/core/test_filters.py
@@ -19,7 +19,7 @@ from polylogue.storage.backends.async_sqlite import SQLiteBackend
 from polylogue.storage.backends.connection import open_connection
 from polylogue.storage.index import rebuild_index
 from polylogue.storage.repository import ConversationRepository
-from tests.infra.helpers import ConversationBuilder
+from tests.infra.storage_records import ConversationBuilder
 
 # =============================================================================
 # TEST DATA: PARAMETRIZATION CONSTANTS (SCREAMING_CASE)
