+++ tests/unit/core/test_filters_adv.py
@@ -24,7 +24,7 @@ from polylogue.storage.backends.async_sqlite import SQLiteBackend
 from polylogue.storage.backends.connection import open_connection
 from polylogue.storage.index import rebuild_index
 from polylogue.storage.repository import ConversationRepository
-from tests.infra.helpers import ConversationBuilder
+from tests.infra.storage_records import ConversationBuilder
 
 
 @pytest.fixture
