+++ tests/unit/storage/test_crud.py
@@ -10,7 +10,7 @@ from polylogue.storage.backends.async_sqlite import SQLiteBackend
 from polylogue.storage.store import (
     ContentBlockRecord,
 )
-from tests.infra.helpers import (
+from tests.infra.storage_records import (
     make_attachment,
     make_conversation,
     make_message,
